split data and fetch functions
