+++ wanikani/client.go
@@ -16,7 +16,26 @@ func GetAPIKey() string {
 	return apiKey
 }
 
-func CreateClient() data.WanikaniClient {
+func CreateClient() WanikaniClient {
 	apiKey := GetAPIKey()
-	return data.WanikaniClient{BaseURL: "https://api.wanikani.com/v2/", APIKey: apiKey, Client: &http.Client{}}
+	return WanikaniClient{BaseURL: "https://api.wanikani.com/v2/", APIKey: apiKey, Client: &http.Client{}}
+}
+
+type WanikaniClient struct {
+	BaseURL string
+	APIKey  string
+	Client  *http.Client
+}
+
+type Client interface {
+	FetchAssignments(levels []string, subjectTypes []string) []data.AssignmentEnvelope
+	FetchWanikaniDataFromEndpoint(endpoint string, data interface{}, parameters map[string]string) error
+	FetchWanikaniDataFromURL(url string, data interface{}) error
+	fetchWanikaniData(request *http.Request, data interface{}) error
+	createAuthorizedRequest(url string) (*http.Request, error)
+	createRequest(endpoint string, parameters map[string]string) (*http.Request, error)
+	convertResponse(response *http.Response, data interface{}) error
+	FetchSpacedRepetitionSystems() []data.SpacedRepetitionSystemEnvelope
+	FetchSubjects(ids []string, levels []string, types []string) []data.SubjectEnvelope
+	FetchUserInformation() data.User
 }
