+++ wanikani/data/spaced_repetition_system.go
@@ -2,8 +2,6 @@ package data
 
 import (
 	"encoding/json"
-	"fmt"
-	"sort"
 )
 
 type Stage struct {
@@ -40,39 +38,4 @@ type SpacedRepetitionSystemEnvelope struct {
 	URL            string                 `json:"url"`
 	DataUploadedAt string                 `json:"data_updated_at"`
 	Data           SpacedRepetitionSystem `json:"data"`
-}
-
-func (o WanikaniClient) FetchSpacedRepetitionSystems() []SpacedRepetitionSystemEnvelope {
-	spacedRepetitionSystemsEnvelope := SpacedRepetitionSystemsEnvelope{}
-
-	err := o.FetchWanikaniDataFromEndpoint("spaced_repetition_systems", &spacedRepetitionSystemsEnvelope, nil)
-	if err != nil {
-		panic(fmt.Errorf("error fetching list of spaced repetition systems: %v", err))
-	}
-
-	return spacedRepetitionSystemsEnvelope.Data
-}
-
-func toIntOrPanic(value json.Number) int64 {
-	if value == "" {
-		return -1
-	}
-	intValue, err := value.Int64()
-	if err != nil {
-		panic(fmt.Errorf("could not convert '%v' to int: %v", value, err))
-	}
-	return intValue
-}
-
-func CreateSpacedRepetitionSystemMap(spacedRepetitionSystemList []SpacedRepetitionSystemEnvelope) map[string]SpacedRepetitionSystem {
-	spacedRepetitionSystems := make(map[string]SpacedRepetitionSystem, len(spacedRepetitionSystemList))
-
-	for _, spacedRepetitionSystem := range spacedRepetitionSystemList {
-		spacedRepetitionSystems[spacedRepetitionSystem.ID.String()] = spacedRepetitionSystem.Data
-		sort.Slice(spacedRepetitionSystem.Data.Stages, func(i, j int) bool {
-			return toIntOrPanic(spacedRepetitionSystem.Data.Stages[i].Interval) < toIntOrPanic(spacedRepetitionSystem.Data.Stages[j].Interval)
-		})
-	}
-
-	return spacedRepetitionSystems
 }
